refactor!: propagate failed assumptions as errors
During workspace creation the code formerly filtered out any workspaces
that didn't have a num property. But this property should always be
there for workspace nodes. The only exception is the scratchpad which is
also a workspace node but doesn't have a number. Propagating this failed
assumption as an error hopefully eases debugging.