refactor!: make a missing focused workspace an error
The code formerly worked around this by pretending a workspace with
number zero is focused. But this case should actually never happen
because there must always be a workspace which has focus. Therefore it
seems better to do an up front check and simplify the implementation of
other methods that rely on the focused workspace.