feat!: limit moving of containers
When moving a container from the last workspace to the next one the old
maximum workspace check was always true. Therefore it was always
possible to move the container and create higher numbered workspaces.
The new implementation checks if the last workspace only contains the
focused container and doesn't perform the move action in this case.