feat: add shifting of following workspace
The new command gives the user the ability to increase the workspace
numbers of all workspaces following the currently focused workspace
by one. The command won't do anything if the focused workspace is the
last one or the next workspace is already empty. The command also tries
to move as few workspaces as possible by closing gaps.