refactor: improve testability of two functions
The logic of the find_all_nodes_by and is_scratchpad_workspace
methods was bound to the swayipc Node struct. This made them hard
to test because creating an instance of this struct requires to
set a lot of fields. The new design uses traits with getter methods
to provide implementations for all types implementing the getter
traits. Furthermore the methods were seperated into different
because they do not share any requirements.

Also added a pipeline step to run the tests.