refactor: split Workspaces creation
Formerly, the Workspaces::new method traversed the node tree, extracted
the required nodes, created wrapper objects, sorted the list and
extracted the focused workspace. That were a lot of different
responsibilities.

The other parts of the struct implementation rely on the sorting and
focused workspace extraction. Due to this coupling those two
responsibilities should still be part of the creation method.

But the other functionality can be easily moved to a free function to
reduce the complexity of the Workspaces::new function.