build: fix the package build for nixos-install
The nixos-install script uses the --store option to build the store at
/mnt/nix/store. Currently crane seems to have a bug when this option is
used. The source filtering removes the Cargo.lock while building the
dependencies. This leads to the following error during build:

    the lock file /build/source/Cargo.lock needs to be updated but --locked was passed to prevent this

This commit explicitly sets the cargoLock option to work around this
bug. After the bug was fixed upstream the option can be removed again.

Furthermore, the GitHub workflow now also checks that a normal Nix
package build as well as one with a different store works.