feat: support named workspaces
Support for named workspaces needed to improvements. The first one
affects the focus and container movement. The command needs to use
the `number` keyword to select a workspace only by its number
instead of its full name. The second change is to preserve the
name part of the full workspace name when shifting the workspace
number.